+++ packages/scripts/config/webpack.config.js
@@ -4,9 +4,8 @@
 const { BundleAnalyzerPlugin } = require( 'webpack-bundle-analyzer' );
 const { CleanWebpackPlugin } = require( 'clean-webpack-plugin' );
 const browserslist = require( 'browserslist' );
-const { sync: glob } = require( 'fast-glob' );
 const MiniCSSExtractPlugin = require( 'mini-css-extract-plugin' );
-const path = require( 'path' );
+const { basename, dirname, resolve } = require( 'path' );
 const ReactRefreshWebpackPlugin = require( '@pmmmwh/react-refresh-webpack-plugin' );
 const TerserPlugin = require( 'terser-webpack-plugin' );
 
@@ -24,6 +23,7 @@ const {
 	hasBabelConfig,
 	hasCssnanoConfig,
 	hasPostCSSConfig,
+	getWebpackEntryPoints,
 } = require( '../utils' );
 
 const isProduction = process.env.NODE_ENV === 'production';
@@ -31,20 +31,6 @@ const mode = isProduction ? 'production' : 'development';
 let target = 'browserslist';
 if ( ! browserslist.findConfig( '.' ) ) {
 	target += ':' + fromConfigRoot( '.browserslistrc' );
-}
-let entry = {};
-if ( process.env.WP_ENTRY ) {
-	entry = JSON.parse( process.env.WP_ENTRY );
-} else {
-	// The script checks whether standard file names can be detected in the `src` folder,
-	// and converts all found files to entry points.
-	const entryFiles = glob( 'src/index.[jt]s?(x)', {
-		absolute: true,
-	} );
-	entryFiles.forEach( ( filepath ) => {
-		const [ entryName ] = path.basename( filepath ).split( '.' );
-		entry[ entryName ] = filepath;
-	} );
 }
 
 const cssLoaders = [
@@ -97,10 +83,10 @@ const cssLoaders = [
 const config = {
 	mode,
 	target,
-	entry,
+	entry: getWebpackEntryPoints(),
 	output: {
 		filename: '[name].js',
-		path: path.resolve( process.cwd(), 'build' ),
+		path: resolve( process.cwd(), 'build' ),
 	},
 	resolve: {
 		alias: {
@@ -118,8 +104,11 @@ const config = {
 					test: /[\\/]style(\.module)?\.(sc|sa|c)ss$/,
 					chunks: 'all',
 					enforce: true,
-					name( module, chunks, cacheGroupKey ) {
-						return `${ cacheGroupKey }-${ chunks[ 0 ].name }`;
+					name( _, chunks, cacheGroupKey ) {
+						const chunkName = chunks[ 0 ].name;
+						return `${ dirname(
+							chunkName
+						) }/${ cacheGroupKey }-${ basename( chunkName ) }`;
 					},
 				},
 				default: false,